Create sampleJSONurlContent
This is what would be pulled from a url if this program ran.  An important part of this whole application is that the program and the url are not in separate domains.  If they are, jsonp would be needed to access cross domains (or some sort of header access control change).  In this case, I'm going to comment out the call to the url because it is located on a database that is only accessible to state of colorado employees.